Added Server-Side Validation (Working)
TODO - Getting the back end to create a new user is proving difficult. Because we're using the supabase/auth-helpers-sveltekit package, and because the service_role key isn't worked into that package, we're not able to use the `supabase.auth.admin` methods, such as the one we need to create a new user. More investigating is needed. We may ultimately need to bring in an entirely new auth module as well as re-examine our database solution (maybe Prisma).