WIP - Working through endpoint refactoring.
Still to do will be re-connecting the fetch requests from the front end to match the response types I'm sending back.

I also need to go back through a few of these and respond with just the naked data, not as an object with a data property. Otherwise I need to res.json() and dig one level for the data property of the response.